Move wasm logic into its own crate
